cleanup, removed getPlatformmVersion
