+++ lib/wearable_sensors_event_channel.dart
@@ -1,6 +1,4 @@
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
-import 'package:wearable_sensors/wearable_sensors.dart';
 
 import 'wearable_sensors_platform_interface.dart';
 
@@ -11,20 +9,6 @@ class EventChannelWearableSensors extends WearableSensorsPlatform {
   final eventChannel = const EventChannel('wearable_sensors');
 
   @override
-  Future<String?> getPlatformVersion() async {
-    final version = "hello this is a version";
-    return version;
-  }
-
-  @override
-  Future<Stream<Map<String, double>>> getSensorStream(MySensorType mySensorType) async {
-    //final version = await methodChannel.invokeMethod<String>('getPlatformVersion');
-    var gyroscopeStream = eventChannel
-            .receiveBroadcastStream(mySensorType.name)
-            .map((dynamic event) => Map<String, double>.from(event));
-    return gyroscopeStream;
-  }
-
   Stream<List<double>> createSensorStream(String channelName) {
     return EventChannel('wearable_sensors/$channelName')
         .receiveBroadcastStream()
